+++ examples/storybook/stories/UploadMedia/Utils/MediaEditor.js
@@ -3,16 +3,22 @@ import { RicosEditor } from 'ricos-editor';
 import { pluginVideo } from 'wix-rich-content-plugin-video';
 import { pluginImage } from 'wix-rich-content-plugin-image';
 import { pluginGallery } from 'wix-rich-content-plugin-gallery';
+import { pluginFileUpload } from 'wix-rich-content-plugin-file-upload';
 import { RichContentEditor } from 'wix-rich-content-editor';
 import PropTypes from 'prop-types';
 
-function getPlugins(handleVideoUpload) {
-  return [pluginImage(), pluginVideo({ handleFileUpload: handleVideoUpload }), pluginGallery()];
+function getPlugins(handleVideoUpload, handleFileUpload) {
+  return [
+    pluginImage(),
+    pluginVideo({ handleFileUpload: handleVideoUpload }),
+    pluginGallery(),
+    pluginFileUpload({ onFileSelected: handleFileUpload }),
+  ];
 }
 
-const MediaEditor = ({ content, handleFileUpload, handleVideoUpload }) => (
-  <RicosEditor plugins={getPlugins(handleVideoUpload)} content={content}>
-    <RichContentEditor helpers={{ handleFileUpload }} />
+const MediaEditor = ({ content, handleFileUpload, handleVideoUpload, handleImageUpload }) => (
+  <RicosEditor plugins={getPlugins(handleVideoUpload, handleFileUpload)} content={content}>
+    <RichContentEditor helpers={{ handleFileSelection: handleImageUpload }} />
   </RicosEditor>
 );
 
@@ -20,6 +26,7 @@ MediaEditor.propTypes = {
   content: PropTypes.object,
   handleFileUpload: PropTypes.func,
   handleVideoUpload: PropTypes.func,
+  handleImageUpload: PropTypes.func,
 };
 
 export default MediaEditor;