+++ examples/storybook/stories/UploadMedia/index.js
@@ -1,4 +1,4 @@
 import { storiesOf } from '@storybook/react';
 import ErrorToastStory from './ErrorToasts';
 
-storiesOf('MediaUpload').add('Error Toasts', ErrorToastStory);
+storiesOf('Media Upload').add('Error Toasts', ErrorToastStory);